Move explore api from helpers to lib and add get likes since method
